+++ lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:sketchspace/canvas/data/worldspace.dart';
 import 'package:sketchspace/classes/demo_debug.dart';
 import 'package:sketchspace/canvas/canvas_context.dart';
-import 'package:sketchspace/canvas/scale.dart';
+import 'package:sketchspace/canvas/data/scale.dart';
 import 'package:sketchspace/classes/settings.dart';
 import 'package:sketchspace/pages/homepage.dart';
 import 'package:flutter/material.dart';
