+++ lib/pages/canvas.dart
@@ -2,9 +2,9 @@ import 'package:sketchspace/brushes/lazy_painter.dart';
 import 'package:sketchspace/canvas/canvas_viewport.dart';
 import 'package:sketchspace/canvas/data/worldspace.dart';
 import 'package:sketchspace/canvas/canvas_context.dart';
-import 'package:sketchspace/canvas/scale.dart';
+import 'package:sketchspace/canvas/data/scale.dart';
 import 'package:sketchspace/classes/settings.dart';
-import 'package:sketchspace/components/canvasUI.dart';
+import 'package:sketchspace/canvas/canvasUI.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:sketchspace/canvas/data/stroke_selector/src/stroke.dart';
